+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All notable changes to this project will be documented here.
 
+<a name="6.0.0"></a>
+## [6.0.0](https://github.com/rawilk/vue-context/releases/6.0.0)
+
+Released 2020-08-03
+
+### Changed
+- **Breaking Change:** Remove named export from entry point ([#60](https://github.com/rawilk/vue-context/issues/60)).
+- Change submenu chevron character code in style sheet.
+
+<a name="5.2.0"></a>
+## [5.2.0](https://github.com/rawilk/vue-context/releases/tag/5.2.0)
+
+Released 2020-05-28
+
+### Added
+- Prop `useScrollHeight` to use element scroll height instead of offset height ([#55](https://github.com/rawilk/vue-context/pull/55)).
+- Prop `useScrollWidth` to use element scroll width instead of offset width ([#55](https://github.com/rawilk/vue-context/pull/55)).
+- Props `heightOffset` and `widthOffset` to specify distance from edge of screen ([#55](https://github.com/rawilk/vue-context/pull/55)). 
+
 <a name="5.1.0"></a>
 ## [5.1.0](https://github.com/rawilk/vue-context/releases/tag/5.1.0)